engine_traits: improve metrics flusher (#7046)
