+++ Cargo.toml
@@ -193,6 +193,7 @@ members = [
   "components/test_util",
   "components/tikv_alloc",
   "components/match_template",
+  "components/engine_traits/tests",
   "components/codec",
   "components/configuration",
   "components/panic_hook",
