+++ components/engine_rocks/src/lib.rs
@@ -61,9 +61,6 @@ pub use compact_listener::*;
 mod properties;
 pub use properties::*;
 
-pub mod metrics_flusher;
-pub use metrics_flusher::*;
-
 pub mod rocks_metrics;
 pub use rocks_metrics::*;
