+++ components/engine_traits/src/engine.rs
@@ -32,6 +32,12 @@ pub trait KvEngine:
     fn snapshot(&self) -> Self::Snapshot;
     fn sync(&self) -> Result<()>;
 
+    /// Flush out metrics. `instance` indicates its name.
+    /// TODO: remove `shared_block_cache`.
+    fn flush_metrics(&self, _instance: &str, _shared_block_cache: bool) {}
+    /// Reset internal statistics.
+    fn reset_statistics(&self) {}
+
     /// This only exists as a temporary hack during refactoring.
     /// It cannot be used forever.
     fn bad_downcast<T: 'static>(&self) -> &T;
