+++ components/engine_traits/tests/Cargo.toml
@@ -1,14 +1,18 @@
 [package]
-name = "engine_traits"
-version = "0.0.1"
+name = "test_engine_traits"
+version = "0.1.0"
 edition = "2018"
 publish = false
 
 [dependencies]
-hex = "0.4"
-protobuf = "2"
-quick-error = "1.2.2"
-tikv_alloc = { path = "../tikv_alloc" }
-tikv_util = { path = "../tikv_util" }
+
+[dev-dependencies]
+engine_rocks = { path = "../../engine_rocks" }
+engine_traits = { path = "../" }
+tempfile = "3.0"
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
+
+[dev-dependencies.rocksdb]
+git = "https://github.com/tikv/rust-rocksdb.git"
+package = "rocksdb"
