+++ components/engine_traits/tests/src/lib.rs
@@ -1,0 +1,6 @@
+#[cfg(test)]
+#[macro_use]
+extern crate slog_global;
+
+#[cfg(test)]
+mod test_metrics_fluster;
